requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cleanup==6.0.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.9.5
python-dotenv==0.19.0
drf-yasg==1.21.4
pyTelegramBotAPI==4.9.0
pytest==7.3.1
pytest-django==4.5.2
django-cors-headers==3.14.0
Faker==18.9.0
pytest-factoryboy==2.5.1
coverage==7.2.5
PyJWT==2.6.0
django-extensions==3.2.3
yookassa==2.5.0